@@ -25,9 +25,8 @@ along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
 	<url>http://maven.apache.org</url>
 
 	<properties>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<java.version>11</java.version>
 		<jboss.logging.version>3.4.1.Final</jboss.logging.version>
 		<keycloak.version>13.0.1</keycloak.version>
 	</properties>
@@ -74,6 +73,15 @@ along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
 
 	<build>
 		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.8.1</version>
+				<configuration>
+					<source>${java.version}</source>
+					<target>${java.version}</target>
+				</configuration>
+			</plugin>
 			<plugin>
 				<artifactId>maven-assembly-plugin</artifactId>
 				<executions>
