@@ -20,7 +20,7 @@ along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
 
 	<groupId>org.shanoir-ng</groupId>
 	<artifactId>shanoir-ng</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
+	<version>2.4.5</version>
 	<packaging>pom</packaging>
 
 	<name>shanoir-ng-parent</name>
@@ -34,8 +34,8 @@ along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
 	</modules>
 
 	<properties>
-		<java.version>1.8</java.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<java.version>11</java.version>
 		<sonar.core.codeCoveragePlugin>jacoco</sonar.core.codeCoveragePlugin>
 		<sonar.jacoco.reportPath>${project.basedir}/../target/jacoco.exec</sonar.jacoco.reportPath>
 		<sonar.language>java</sonar.language>
