@@ -5,29 +5,32 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
 import java.text.ParseException;
+import java.util.ArrayList;
 import java.util.Date;
-
 import javax.swing.JButton;
 import javax.swing.JOptionPane;
-import javax.xml.datatype.XMLGregorianCalendar;
 
 import org.apache.log4j.Logger;
-import org.shanoir.dicom.importer.PreImportData;
 import org.shanoir.dicom.importer.UploadJob;
 import org.shanoir.uploader.ShUpConfig;
 import org.shanoir.uploader.ShUpOnloadConfig;
 import org.shanoir.uploader.gui.ImportDialog;
 import org.shanoir.uploader.gui.MainWindow;
-import org.shanoir.uploader.model.Center;
-import org.shanoir.uploader.model.Investigator;
-import org.shanoir.uploader.model.Study;
-import org.shanoir.uploader.model.StudyCard;
-import org.shanoir.uploader.model.dto.ExaminationDTO;
-import org.shanoir.uploader.service.soap.ShanoirUploaderServiceClient;
+import org.shanoir.uploader.model.rest.Examination;
+import org.shanoir.uploader.model.rest.HemisphericDominance;
+import org.shanoir.uploader.model.rest.IdName;
+import org.shanoir.uploader.model.rest.ImagedObjectCategory;
+import org.shanoir.uploader.model.rest.PseudonymusHashValues;
+import org.shanoir.uploader.model.rest.Sex;
+import org.shanoir.uploader.model.rest.Study;
+import org.shanoir.uploader.model.rest.StudyCard;
+import org.shanoir.uploader.model.rest.Subject;
+import org.shanoir.uploader.model.rest.SubjectStudy;
+import org.shanoir.uploader.model.rest.SubjectType;
+import org.shanoir.uploader.model.rest.importer.ImportJob;
+import org.shanoir.uploader.service.rest.ShanoirUploaderServiceClient;
+import org.shanoir.uploader.utils.ImportUtils;
 import org.shanoir.uploader.utils.Util;
-import org.shanoir.ws.generated.uploader.PseudonymusHashValues;
-import org.shanoir.ws.generated.uploader.SubjectDTO;
-import org.shanoir.ws.generated.uploader.SubjectStudyDTO;
 
 /**
  * This class implements the logic when the start import button is clicked.
@@ -45,19 +48,19 @@ public class ImportFinishActionListener implements ActionListener {
 	
 	private File uploadFolder;
 	
-	private org.shanoir.uploader.model.dto.SubjectDTO subjectDTO;
+	private Subject subject;
 	
 	private ShanoirUploaderServiceClient shanoirUploaderServiceClient;
 	
-	private ImportStudyAndStudyCardCBItemListener importStudyAndStudyCardCBIL;
+	private ImportStudyAndStudyCardCBItemListener importStudyAndStudyCardCBILNG;
 
-	public ImportFinishActionListener(final MainWindow mainWindow, UploadJob uploadJob, File uploadFolder, org.shanoir.uploader.model.dto.SubjectDTO subjectDTO,
-			ImportStudyAndStudyCardCBItemListener importStudyAndStudyCardCBIL) {
+	public ImportFinishActionListener(final MainWindow mainWindow, UploadJob uploadJob, File uploadFolder, Subject subject,
+			ImportStudyAndStudyCardCBItemListener importStudyAndStudyCardCBILNG) {
 		this.mainWindow = mainWindow;
 		this.uploadJob = uploadJob;
 		this.uploadFolder = uploadFolder;
-		this.subjectDTO = subjectDTO;
-		this.importStudyAndStudyCardCBIL = importStudyAndStudyCardCBIL;
+		this.subject = subject;
+		this.importStudyAndStudyCardCBILNG = importStudyAndStudyCardCBILNG;
 		this.shanoirUploaderServiceClient = ShUpOnloadConfig.getShanoirUploaderServiceClient();
 	}
 
@@ -65,17 +68,18 @@ public class ImportFinishActionListener implements ActionListener {
 	 * This method contains all the logic which is performed when the start import
 	 * button is clicked.
 	 */
+	@Override
 	public void actionPerformed(final ActionEvent event) {
 		final Study study = (Study) mainWindow.importDialog.studyCB.getSelectedItem();
 		final StudyCard studyCard = (StudyCard) mainWindow.importDialog.studyCardCB.getSelectedItem();
-		if (study == null || study.getId() == null || studyCard == null || studyCard.getId() == null) {
+		if (study == null || study.getId() == null || studyCard == null || studyCard.getName() == null) {
 			return;
 		}
 		if (ShUpConfig.isModeSubjectCommonNameManual()) {
 			// minimal length for subject common name is 2, same for subject study identifier
 			if (mainWindow.importDialog.subjectTextField.getText().length() < 2
-				|| (!mainWindow.importDialog.subjectStudyIdentifierTF.getText().isEmpty()
-						&& mainWindow.importDialog.subjectStudyIdentifierTF.getText().length() < 2)) {
+				|| !mainWindow.importDialog.subjectStudyIdentifierTF.getText().isEmpty()
+						&& mainWindow.importDialog.subjectStudyIdentifierTF.getText().length() < 2) {
 				JOptionPane.showMessageDialog(mainWindow.frame,
 						mainWindow.resourceBundle.getString("shanoir.uploader.systemErrorDialog.error.subject.creation"),
 						"Error", JOptionPane.ERROR_MESSAGE);
@@ -87,15 +91,12 @@ public class ImportFinishActionListener implements ActionListener {
 		((JButton) event.getSource()).setEnabled(false);
 		mainWindow.setCursor(Cursor.getPredefinedCursor(Cursor.WAIT_CURSOR));
 
-		// TODO the next lines are totally strange as GUI and two different DTOs are used: totally strange,
-		// but I did not refactor here as too time demanding at the moment and bad legacy of developer before
-		Long subjectId = null;
-		String subjectName = null;
-		// No existing subject found
-		if (subjectDTO == null) {
-			SubjectDTO subjectDTOWSDL = null;
+		/**
+		 * Handle subject here: creation or use existing
+		 */
+		if (subject == null) {
 			try {
-				 subjectDTOWSDL = fillSubjectDTO(mainWindow.importDialog, uploadJob, study.getId());
+				 subject = fillSubject(mainWindow.importDialog, uploadJob);
 			} catch (ParseException e) {
 				logger.error(e.getMessage(), e);
 				JOptionPane.showMessageDialog(mainWindow.frame,
@@ -105,8 +106,14 @@ public class ImportFinishActionListener implements ActionListener {
 				((JButton) event.getSource()).setEnabled(true);
 				return;
 			}
-			subjectDTOWSDL = shanoirUploaderServiceClient.createSubject(studyCard.getId(), ShUpConfig.isModeSubjectCommonNameManual(), subjectDTOWSDL);
-			if (subjectDTOWSDL == null) {
+			ImportUtils.addSubjectStudy(study, subject,
+					(SubjectType) mainWindow.importDialog.subjectTypeCB.getSelectedItem(),
+					mainWindow.importDialog.subjectIsPhysicallyInvolvedCB.isSelected(),
+					mainWindow.importDialog.subjectStudyIdentifierTF.getText());
+			// create subject with subject-study filled to avoid access denied exception because of rights check
+			Long centerId = studyCard.getAcquisitionEquipment().getCenter().getId();
+			subject = shanoirUploaderServiceClient.createSubject(subject, ShUpConfig.isModeSubjectCommonNameManual(), centerId);
+			if (subject == null) {
 				JOptionPane.showMessageDialog(mainWindow.frame,
 						mainWindow.resourceBundle.getString("shanoir.uploader.systemErrorDialog.error.wsdl.subjectcreator.createSubjectFromShup"),
 						"Error", JOptionPane.ERROR_MESSAGE);
@@ -114,61 +121,74 @@ public class ImportFinishActionListener implements ActionListener {
 				((JButton) event.getSource()).setEnabled(true);
 				return;
 			} else {
-				subjectId = new Long(subjectDTOWSDL.getId());
-				subjectName = subjectDTOWSDL.getName();
-				handleSubjectStudy(study, subjectId);
-				logger.info("Auto-Import: subject created on server with ID: " + subjectId);
+				logger.info("Auto-Import: subject created on server with ID: " + subject.getId());
 			}
-		// Existing subject found: but check if already in selected study (RelSubjectStudy)
 		} else {
-			subjectId = subjectDTO.getId();
-			subjectName = subjectDTO.getName();
-			handleSubjectStudy(study, subjectId);
-			logger.info("Auto-Import: subject used on server with ID: " + subjectId);
-		}
-		Long examinationId = null;
-		if (mainWindow.importDialog.mrExaminationNewExamCB.isSelected()) {
-			Center center = (Center) mainWindow.importDialog.mrExaminationCenterCB.getSelectedItem();
-			Investigator investigator = (Investigator) mainWindow.importDialog.mrExaminationExamExecutiveCB.getSelectedItem();
-			Date examinationDate = (Date) mainWindow.importDialog.mrExaminationDateDP.getModel().getValue();
-			String examinationComment = mainWindow.importDialog.mrExaminationCommentTF.getText();
-			if (investigator != null && center != null && examinationDate != null) {
-				long createExaminationId = shanoirUploaderServiceClient.createExamination(study.getId(), subjectId, new Long(center.getId()),
-						new Long(investigator.getId()), examinationDate, examinationComment);
-				if (createExaminationId == -1) {
+			// if rel-subject-study does not exist for existing subject, create one
+			if (importStudyAndStudyCardCBILNG.getSubjectStudy() == null) {
+				ImportUtils.addSubjectStudy(study, subject,
+						(SubjectType) mainWindow.importDialog.subjectTypeCB.getSelectedItem(),
+						mainWindow.importDialog.subjectIsPhysicallyInvolvedCB.isSelected(),
+						mainWindow.importDialog.subjectStudyIdentifierTF.getText());
+				if (shanoirUploaderServiceClient.createSubjectStudy(subject) == null) {
 					JOptionPane.showMessageDialog(mainWindow.frame,
-							mainWindow.resourceBundle.getString("shanoir.uploader.systemErrorDialog.error.wsdl.createmrexamination"),
+							mainWindow.resourceBundle.getString("shanoir.uploader.systemErrorDialog.error.wsdl.subjectcreator.createSubjectFromShup"),
 							"Error", JOptionPane.ERROR_MESSAGE);
 					mainWindow.setCursor(null); // turn off the wait cursor
 					((JButton) event.getSource()).setEnabled(true);
 					return;
-				} else {
-					examinationId = new Long(createExaminationId);
-					logger.info("Auto-Import: examination created on server with ID: " + examinationId);
 				}
-			} else {
+				
+			}
+			logger.info("Auto-Import: subject used on server with ID: " + subject.getId());
+		}
+		
+		Long examinationId = null;
+		if (mainWindow.importDialog.mrExaminationNewExamCB.isSelected()) {
+			Examination examinationDTO = new Examination();
+			examinationDTO.setStudyId(study.getId());
+			examinationDTO.setSubjectId(subject.getId());
+			IdName center = (IdName) mainWindow.importDialog.mrExaminationCenterCB.getSelectedItem();
+			examinationDTO.setCenterId(center.getId());
+//			Investigator investigator = (Investigator) mainWindow.importDialog.mrExaminationExamExecutiveCB.getSelectedItem();
+			Date examinationDate = (Date) mainWindow.importDialog.mrExaminationDateDP.getModel().getValue();
+			String examinationComment = mainWindow.importDialog.mrExaminationCommentTF.getText();
+			examinationDTO.setExaminationDate(examinationDate);
+			examinationDTO.setComment(examinationComment);
+
+			/**
+			 * TODO handle investigators here or decide finally to delete them in sh-ng
+			 */
+			examinationDTO = shanoirUploaderServiceClient.createExamination(examinationDTO);
+			if (examinationDTO == null) {
 				JOptionPane.showMessageDialog(mainWindow.frame,
 						mainWindow.resourceBundle.getString("shanoir.uploader.systemErrorDialog.error.wsdl.createmrexamination"),
 						"Error", JOptionPane.ERROR_MESSAGE);
 				mainWindow.setCursor(null); // turn off the wait cursor
 				((JButton) event.getSource()).setEnabled(true);
 				return;
+			} else {
+				examinationId = examinationDTO.getId();
+				logger.info("Auto-Import: examination created on server with ID: " + examinationId);
 			}
 		} else {
-			ExaminationDTO examinationDTO = (ExaminationDTO) mainWindow.importDialog.mrExaminationExistingExamCB.getSelectedItem();
+			Examination examinationDTO = (Examination) mainWindow.importDialog.mrExaminationExistingExamCB.getSelectedItem();
 			examinationId = examinationDTO.getId();
 			logger.info("Auto-Import: examination used on server with ID: " + examinationId);
 		}
-		
+				
 		/**
-		 * 3. Fill PreImportData, later added to upload-job.xml
+		 * 3. Fill import-job.json
 		 */
-		PreImportData preImportData = fillPreImportData(mainWindow.importDialog, subjectId, examinationId);
-		Runnable runnable = new ImportFinishRunnable(uploadJob, uploadFolder, preImportData, subjectName);
+		//Exchange exchange = prepareExchange(mainWindow.importDialog, subject.getName(), subject.getId(), examinationId);
+		ImportJob importJob = ImportUtils.prepareImportJob(uploadJob, subject.getName(), subject.getId(), examinationId, (Study) mainWindow.importDialog.studyCB.getSelectedItem(), (StudyCard) mainWindow.importDialog.studyCardCB.getSelectedItem());
+		Runnable runnable = new ImportFinishRunnable(uploadJob, uploadFolder, importJob, subject.getName());
 		Thread thread = new Thread(runnable);
 		thread.start();
 		
 		mainWindow.importDialog.setVisible(false);
+		mainWindow.importDialog.mrExaminationExamExecutiveLabel.setVisible(true);
+		mainWindow.importDialog.mrExaminationExamExecutiveCB.setVisible(true);
 		mainWindow.setCursor(null); // turn off the wait cursor
 		((JButton) event.getSource()).setEnabled(true);
 		
@@ -176,53 +196,49 @@ public class ImportFinishActionListener implements ActionListener {
 				ShUpConfig.resourceBundle.getString("shanoir.uploader.import.start.auto.import.message"),
 				"Import", JOptionPane.INFORMATION_MESSAGE);
 	}
+	
 
-	private void handleSubjectStudy(final Study study, final Long subjectId) {
-		if (importStudyAndStudyCardCBIL.getSubjectStudyDTO() == null) {
-			SubjectStudyDTO subjectStudyDTO = new SubjectStudyDTO();
-			subjectStudyDTO.setStudyId(study.getId());
-			subjectStudyDTO.setSubjectId(subjectId);
-			subjectStudyDTO.setSubjectStudyIdentifier(mainWindow.importDialog.subjectStudyIdentifierTF.getText());
-			subjectStudyDTO.setSubjectType((String) mainWindow.importDialog.subjectTypeCB.getSelectedItem());
-			subjectStudyDTO.setPhysicallyInvolved(mainWindow.importDialog.subjectIsPhysicallyInvolvedCB.isSelected());
-			subjectStudyDTO = shanoirUploaderServiceClient.createSubjectStudy(subjectStudyDTO);
-			logger.info("Auto-import: RelSubjectStudy created with id: " + subjectStudyDTO.getId() + " for subject: " + subjectId + " in study: " + study.getName());
-		}
-	}
 
-	/**
-	 * This method fills the SubjectDTO and the RelSubjectStudy that will be send to the web service
-	 * to create a new subject and integrate it into a study.
-	 * 
-	 * @param importDialog
-	 * @param dicomData
-	 * @return
-	 * @throws ParseException 
-	 */
-	private SubjectDTO fillSubjectDTO(final ImportDialog importDialog, final UploadJob uploadJob, final Long studyId) throws ParseException {
-		final SubjectDTO subjectDTO = new SubjectDTO();
-		if (ShUpConfig.isModeSubjectCommonNameManual()) {
-			subjectDTO.setName(importDialog.subjectTextField.getText());
-		}
+	private Subject fillSubject(final ImportDialog importDialog, final UploadJob uploadJob) throws ParseException {
+		final Subject subjectDTO = new Subject();
+		/**
+		 * Values coming from UploadJob
+		 */
 		subjectDTO.setIdentifier(uploadJob.getSubjectIdentifier());
         Date birthDate = ShUpConfig.formatter.parse(uploadJob.getPatientBirthDate());
-		XMLGregorianCalendar birthDateXMLGC = Util.toXMLGregorianCalendar(birthDate);
-		subjectDTO.setBirthDate(birthDateXMLGC);
-		subjectDTO.setSex(uploadJob.getPatientSex());
-		subjectDTO.setImagedObjectCategory((String) importDialog.subjectImageObjectCategoryCB.getSelectedItem());
-		subjectDTO.setLanguageHemisphericDominance((String) importDialog.subjectLanguageHemisphericDominanceCB.getSelectedItem());
-		subjectDTO.setManualHemisphericDominance((String) importDialog.subjectManualHemisphericDominanceCB.getSelectedItem());
+		subjectDTO.setBirthDate(Util.convertToLocalDateViaInstant(birthDate));
+		if (uploadJob.getPatientSex().compareTo(Sex.F.name()) == 0) {
+			subjectDTO.setSex(Sex.F);
+		} else if (uploadJob.getPatientSex().compareTo(Sex.M.name()) == 0) {
+			subjectDTO.setSex(Sex.M);
+		}
 		if (ShUpConfig.isModePseudonymus()) {
 			fillPseudonymusHashValues(uploadJob, subjectDTO);
 		}
+		/**
+		 * Values coming from ImportDialog
+		 */
+		if (ShUpConfig.isModeSubjectCommonNameManual()) {
+			subjectDTO.setName(importDialog.subjectTextField.getText());
+		}
+		subjectDTO.setImagedObjectCategory((ImagedObjectCategory) importDialog.subjectImageObjectCategoryCB.getSelectedItem());
+		String languageHemDom = (String) importDialog.subjectLanguageHemisphericDominanceCB.getSelectedItem();
+		if (HemisphericDominance.Left.getName().compareTo(languageHemDom) == 0) {
+			subjectDTO.setLanguageHemisphericDominance(HemisphericDominance.Left);
+		} else if (HemisphericDominance.Right.getName().compareTo(languageHemDom) == 0) {
+			subjectDTO.setLanguageHemisphericDominance(HemisphericDominance.Right);
+		}
+		String manualHemDom = (String) importDialog.subjectManualHemisphericDominanceCB.getSelectedItem();
+		if (HemisphericDominance.Left.getName().compareTo(manualHemDom) == 0) {
+			subjectDTO.setManualHemisphericDominance(HemisphericDominance.Left);
+		} else if (HemisphericDominance.Right.getName().compareTo(manualHemDom) == 0) {
+			subjectDTO.setManualHemisphericDominance(HemisphericDominance.Right);
+		}
+		subjectDTO.setSubjectStudyList(new ArrayList<SubjectStudy>());
 		return subjectDTO;
 	}
 
-	/**
-	 * @param dicomData
-	 * @param subjectDTO
-	 */
-	private void fillPseudonymusHashValues(final UploadJob uploadJob, final SubjectDTO subjectDTO) {
+	private void fillPseudonymusHashValues(final UploadJob uploadJob, final Subject subjectDTO) {
 		PseudonymusHashValues pseudonymusHashValues = new PseudonymusHashValues();
 		pseudonymusHashValues.setFirstNameHash1(uploadJob.getFirstNameHash1());
 		pseudonymusHashValues.setFirstNameHash2(uploadJob.getFirstNameHash2());
@@ -235,27 +251,6 @@ public class ImportFinishActionListener implements ActionListener {
 		pseudonymusHashValues.setBirthNameHash3(uploadJob.getBirthNameHash3());
 		pseudonymusHashValues.setBirthDateHash(uploadJob.getBirthDateHash());
 		subjectDTO.setPseudonymusHashValues(pseudonymusHashValues);
-	}
-
-	private PreImportData fillPreImportData(ImportDialog importDialog, Long subjectId, Long examinationId) {
-		PreImportData preImportData = new PreImportData();
-		preImportData.setAutoImportEnable(true);
-		
-		org.shanoir.dicom.importer.Study exportStudy = new org.shanoir.dicom.importer.Study();
-		Study study = (Study) importDialog.studyCB.getSelectedItem();
-		exportStudy.setId(study.getId().intValue());
-		exportStudy.setName(study.getName());
-		preImportData.setStudy(exportStudy);
-		
-		org.shanoir.dicom.importer.StudyCard exportStudyCard = new org.shanoir.dicom.importer.StudyCard();
-		StudyCard studyCard = (StudyCard) importDialog.studyCardCB.getSelectedItem();
-		exportStudyCard.setId(studyCard.getId().intValue());
-		exportStudyCard.setName(studyCard.getName());
-		preImportData.setStudycard(exportStudyCard);
-		
-		preImportData.setSubjectId(subjectId);
-		preImportData.setExaminationId(examinationId);
-		return preImportData;
 	}
 
 }
